inline isPackageManifest
